requirements.txt
absl-py==1.4.0
ale-py==0.8.1
apache-beam==2.48.0
appdirs==1.4.4
appnope==0.1.3
array-record==0.2.0
asttokens==2.2.1
astunparse==1.6.3
attrs==23.1.0
audioread==3.0.0
backcall==0.2.0
blinker==1.6.2
blosc2==2.0.0
bokeh==3.1.1
bz2file==0.98
cached-property==1.5.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
chex==0.1.7
click==8.1.3
cloudpickle==2.2.1
cmake==3.26.3
contourpy==1.0.7
crcmod==1.7
cycler==0.11.0
Cython==0.29.35
decorator==5.1.1
dill==0.3.1.1
dm-sonnet==2.0.1
dm-tree==0.1.8
dnspython==2.3.0
docopt==0.6.2
dopamine-rl==3.2.1
etils==1.3.0
executing==1.2.0
fastavro==1.7.4
fasteners==0.18
Flask==2.3.2
flatbuffers==23.5.26
flax==0.6.10
fonttools==4.39.4
future==0.18.3
gast==0.4.0
gevent==22.10.2
gin-config==0.5.0
google-api-core==2.11.0
google-api-python-client==2.88.0
google-apitools==0.5.31
google-auth==2.19.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-bigquery==3.11.0
google-cloud-bigquery-storage==2.20.0
google-cloud-bigtable==2.17.0
google-cloud-core==2.3.2
google-cloud-datastore==2.15.2
google-cloud-dlp==3.12.1
google-cloud-language==2.10.0
google-cloud-pubsub==2.17.1
google-cloud-pubsublite==1.8.2
google-cloud-recommendations-ai==0.10.3
google-cloud-spanner==3.35.1
google-cloud-videointelligence==2.11.2
google-cloud-vision==3.4.2
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpc-google-iam-v1==0.12.6
grpcio==1.54.2
grpcio-status==1.54.2
gunicorn==20.1.0
gym==0.26.2
gym-notices==0.0.8
h5py==3.8.0
hdfs==2.7.0
httplib2==0.22.0
idna==3.4
imageio==2.30.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
intervaltree==3.1.0
ipython==8.13.2
itsdangerous==2.1.2
jax==0.4.11
jaxlib==0.4.11
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpickle==3.0.1
keras==2.13.1rc0
keras-tuner==1.3.5
kfac==0.2.0
kiwisolver==1.4.4
kt-legacy==1.0.5
lazy_loader==0.2
libclang==16.0.0
librosa==0.10.0.post2
llvmlite==0.40.0
magenta==2.0.1
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mesh-tensorflow==0.1.21
meson==1.1.1
MIDIUtil==1.2.1
mido==1.2.6
mir-eval==0.7
ml-dtypes==0.1.0
more-itertools==9.1.0
mpmath==1.3.0
msgpack==1.0.5
music21==8.3.0
ndindex==1.7
nest-asyncio==1.5.6
networkx==3.1
ninja==1.11.1
note-seq==0.0.5
numba==0.57.0
numexpr==2.8.4
numpy==1.23.1
oauth2client==4.1.3
oauthlib==3.2.2
objsize==0.6.1
opencv-python==4.7.0.72
opt-einsum==3.3.0
optax==0.1.5
orbax-checkpoint==0.2.4
orjson==3.9.0
overrides==6.5.0
packaging==23.1
pandas==2.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pooch==1.6.0
pretty-midi==0.2.10
promise==2.3
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.23.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydot==1.4.2
pydub==0.25.1
pyFluidSynth==1.3.2
pygame==2.4.0
Pygments==2.15.1
pygtrie==2.5.0
pymongo==4.3.3
pyparsing==3.0.9
pypng==0.20220715.0
python-dateutil==2.8.2
python-rtmidi==1.1.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.1
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
seaborn==0.12.2
six==1.16.0
sk-video==1.1.10
sortedcontainers==2.4.0
soundfile==0.12.1
sox==1.4.1
soxr==0.3.5
sqlparse==0.4.4
stack-data==0.6.2
sympy==1.12
tables==3.8.0
tabulate==0.9.0
tensor2tensor==1.15.7
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tensorflow==2.13.0rc1
tensorflow-addons==0.20.0
tensorflow-datasets==4.9.2
tensorflow-estimator==2.13.0rc0
tensorflow-gan==2.1.0
tensorflow-hub==0.13.0
tensorflow-macos==2.13.0rc1
tensorflow-metadata==1.13.1
tensorflow-probability==0.7.0
tensorstore==0.1.37
termcolor==2.3.0
tf-slim==1.1.0
threadpoolctl==3.1.0
tifffile==2023.4.12
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typeguard==2.13.3
typing_extensions==4.6.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
wcwidth==0.2.6
webcolors==1.13
Werkzeug==2.3.4
wrapt==1.15.0
xyzservices==2023.5.0
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
zstandard==0.21.0
